Fake ekuiper/requirements.txt
psycopg2-binary==2.9.9
argparse
paho-mqtt==2.1.0